[playground.spaces.cg.fem-localfunctions] moved ContinousLagrange -> CG
